requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# For faster fine-tuning (optional but recommended)
# unsloth @ git+https://github.com/unslothai/unsloth.git

# Embeddings and ML utilities
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# LLM API
openai>=1.0.0
anthropic>=0.7.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
